+++ indispensable-cosef/src/commonMain/kotlin/at/asitplus/signum/indispensable/cosef/CoseKeyParams.kt
@@ -5,6 +5,7 @@ import at.asitplus.KmmResult.Companion.failure
 import at.asitplus.catching
 import at.asitplus.signum.indispensable.CryptoPublicKey
 import at.asitplus.signum.indispensable.SpecializedCryptoPublicKey
+import at.asitplus.signum.indispensable.asn1.Asn1Integer
 import at.asitplus.signum.indispensable.asn1.encoding.decodeFromAsn1ContentBytes
 
 /**
@@ -163,10 +164,10 @@ sealed class CoseKeyParams : SpecializedCryptoPublicKey {
 
         override fun toCryptoPublicKey(): KmmResult<CryptoPublicKey> = catching {
             CryptoPublicKey.RSA(
-                n = n ?: throw IllegalArgumentException("Missing modulus n"),
-                e = Int.decodeFromAsn1ContentBytes(e ?:
-                    throw IllegalArgumentException("Missing or invalid exponent e"))
-            )
+                n = Asn1Integer.fromUnsignedByteArray(
+                    n ?: throw IllegalArgumentException("Missing modulus n")),
+                e = Asn1Integer.fromUnsignedByteArray(
+                    e ?: throw IllegalArgumentException("Missing or invalid exponent e")))
         }
     }