+++ indispensable-cosef/src/commonTest/kotlin/at/asitplus/signum/indispensable/cosef/CoseEqualsTest.kt
@@ -1,5 +1,7 @@
 package at.asitplus.signum.indispensable.cosef
 
+import at.asitplus.signum.indispensable.CryptoSignature
+import at.asitplus.signum.indispensable.cosef.io.Base16Strict
 import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapper
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
@@ -9,65 +11,120 @@ import io.kotest.property.arbitrary.byte
 import io.kotest.property.arbitrary.byteArray
 import io.kotest.property.arbitrary.int
 import io.kotest.property.checkAll
+import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
+import kotlinx.serialization.Serializable
 
 class CoseEqualsTest : FreeSpec({
-    "Test Equals" - {
-        checkAll(
-            Arb.byteArray(
-                length = Arb.int(0, 10),
-                content = Arb.byte()
-            )
-        ) { s1 ->
-            val signed1 = CoseSigned(
-                protectedHeader = ByteStringWrapper<CoseHeader>(CoseHeader()),
+    "equals with byte array" {
+        checkAll(Arb.byteArray(length = Arb.int(0, 10), content = Arb.byte())) { bytes ->
+            val bytesSigned1 = CoseSigned<ByteArray>(
+                protectedHeader = ByteStringWrapper(CoseHeader()),
                 unprotectedHeader = null,
-                payload = s1,
-                rawSignature = s1
+                payload = bytes,
+                rawSignature = bytes
             )
-            val signed11 = CoseSigned(
-                protectedHeader = ByteStringWrapper<CoseHeader>(CoseHeader()),
+            val bytesSigned2 = CoseSigned<ByteArray>(
+                protectedHeader = ByteStringWrapper(CoseHeader()),
                 unprotectedHeader = null,
-                payload = s1,
-                rawSignature = s1
+                payload = bytes,
+                rawSignature = bytes
             )
 
-            signed1 shouldBe signed1
-            signed11 shouldBe signed1
-            signed1.hashCode() shouldBe signed1.hashCode()
-            signed1.hashCode() shouldBe signed11.hashCode()
+            bytesSigned1 shouldBe bytesSigned1
+            bytesSigned2 shouldBe bytesSigned1
+            bytesSigned1.hashCode() shouldBe bytesSigned1.hashCode()
+            bytesSigned1.hashCode() shouldBe bytesSigned2.hashCode()
 
-            val s2 = s1.reversedArray().let { it + it + 1 + 3 + 5 }
-            val signed2 = CoseSigned(
-                protectedHeader = ByteStringWrapper<CoseHeader>(CoseHeader()),
+            val reversed = bytes.reversedArray().let { it + it + 1 + 3 + 5 }
+            val reversedSigned1 = CoseSigned<ByteArray>(
+                protectedHeader = ByteStringWrapper(CoseHeader()),
                 unprotectedHeader = null,
-                payload = s2,
-                rawSignature = s2
+                payload = reversed,
+                rawSignature = reversed
             )
-            val signed22 = CoseSigned(
-                protectedHeader = ByteStringWrapper<CoseHeader>(CoseHeader()),
+            val reversedSigned2 = CoseSigned<ByteArray>(
+                protectedHeader = ByteStringWrapper(CoseHeader()),
                 unprotectedHeader = null,
-                payload = s2,
-                rawSignature = s2
+                payload = reversed,
+                rawSignature = reversed
             )
 
-            signed22 shouldBe signed22
-            signed22 shouldBe signed2
+            reversedSigned2 shouldBe reversedSigned2
+            reversedSigned2 shouldBe reversedSigned1
 
-            signed2.hashCode() shouldBe signed2.hashCode()
-            signed2.hashCode() shouldBe signed22.hashCode()
+            reversedSigned1.hashCode() shouldBe reversedSigned1.hashCode()
+            reversedSigned1.hashCode() shouldBe reversedSigned2.hashCode()
 
-            signed1 shouldNotBe signed2
-            signed1 shouldNotBe signed22
+            bytesSigned1 shouldNotBe reversedSigned1
+            bytesSigned1 shouldNotBe reversedSigned2
 
-            signed1.hashCode() shouldNotBe signed2.hashCode()
-            signed1.hashCode() shouldNotBe signed22.hashCode()
+            bytesSigned1.hashCode() shouldNotBe reversedSigned1.hashCode()
+            bytesSigned1.hashCode() shouldNotBe reversedSigned2.hashCode()
 
-            signed2 shouldNotBe signed1
-            signed2 shouldNotBe signed11
+            reversedSigned1 shouldNotBe bytesSigned1
+            reversedSigned1 shouldNotBe bytesSigned2
 
-            signed2.hashCode() shouldNotBe signed1.hashCode()
-            signed2.hashCode() shouldNotBe signed11.hashCode()
+            reversedSigned1.hashCode() shouldNotBe bytesSigned1.hashCode()
+            reversedSigned1.hashCode() shouldNotBe bytesSigned2.hashCode()
+        }
+    }
+
+    "equals with data class" {
+        checkAll(Arb.byteArray(length = Arb.int(0, 10), content = Arb.byte())) { bytes ->
+            val payload = DataClass(content = bytes.encodeToString(Base16Strict))
+            val bytesSigned1 = CoseSigned.fromObject<DataClass>(
+                protectedHeader = CoseHeader(),
+                unprotectedHeader = null,
+                payload = payload,
+                signature = CryptoSignature.RSAorHMAC(bytes)
+            )
+            val bytesSigned2 = CoseSigned.fromObject<DataClass>(
+                protectedHeader = CoseHeader(),
+                unprotectedHeader = null,
+                payload = payload,
+                signature = CryptoSignature.RSAorHMAC(bytes)
+            )
+
+            bytesSigned1 shouldBe bytesSigned1
+            bytesSigned2 shouldBe bytesSigned1
+            bytesSigned1.hashCode() shouldBe bytesSigned1.hashCode()
+            bytesSigned1.hashCode() shouldBe bytesSigned2.hashCode()
+
+            val reversed = DataClass(content = bytes.reversedArray().let { it + it + 1 + 3 + 5 }.encodeToString(Base16Strict))
+            val reversedSigned1 = CoseSigned.fromObject<DataClass>(
+                protectedHeader = CoseHeader(),
+                unprotectedHeader = null,
+                payload = reversed,
+                signature = CryptoSignature.RSAorHMAC(bytes)
+            )
+            val reversedSigned2 = CoseSigned.fromObject<DataClass>(
+                protectedHeader = CoseHeader(),
+                unprotectedHeader = null,
+                payload = reversed,
+                signature = CryptoSignature.RSAorHMAC(bytes)
+            ).also { println(it.serialize().encodeToString(Base16Strict))}
+
+            reversedSigned2 shouldBe reversedSigned2
+            reversedSigned2 shouldBe reversedSigned1
+
+            reversedSigned1.hashCode() shouldBe reversedSigned1.hashCode()
+            reversedSigned1.hashCode() shouldBe reversedSigned2.hashCode()
+
+            bytesSigned1 shouldNotBe reversedSigned1
+            bytesSigned1 shouldNotBe reversedSigned2
+
+            bytesSigned1.hashCode() shouldNotBe reversedSigned1.hashCode()
+            bytesSigned1.hashCode() shouldNotBe reversedSigned2.hashCode()
+
+            reversedSigned1 shouldNotBe bytesSigned1
+            reversedSigned1 shouldNotBe bytesSigned2
+
+            reversedSigned1.hashCode() shouldNotBe bytesSigned1.hashCode()
+            reversedSigned1.hashCode() shouldNotBe bytesSigned2.hashCode()
         }
 
     }
 })
+
+@Serializable
+data class DataClass(val content: String)